updating version 1.0.03

+++ src/objectIO/ObjectReader.java
@@ -84,7 +84,8 @@ public class ObjectReader {
 				case("int"): if (f.getInt(o)           == NON_INITIALIZED_INT)     b = false; break;
 				case("Integer"): if (((int) f.getInt(o))         == NON_INITIALIZED_INT)     b = false; break;
 				case("double"):  if (f.getDouble(o)    == NON_INITIALIZED_DOUBLE)  b = false; break;
-				case("Double"):  if (((double) f.getDouble(o))    == NON_INITIALIZED_DOUBLE)  b = false; break;
+				case("Double"):  if (f.getDouble(o)    == NON_INITIALIZED_DOUBLE)  b = false; break;
+//				case("Double"):  if (((double) f.getDouble(o))    == NON_INITIALIZED_DOUBLE)  b = false; break;
 				case("boolean"): if (f.get(o)          == null)                    b = false; break;
 				case("String"):  if ((String) f.get(o) == NON_INITIALIZED_STRING)  b = false; break;
 				}
